// types/types.ts
export interface UserProfile {
  name: string
  age: string
  jobTitle: string
  salary: string
}

export interface Expenses {
  rent: string
  utilities: string
  food: string
  transportation: string
  entertainment: string
}

export interface FinancialStatus {
  savings: number
  investments: number
  debt: number
  netWorth: number
}